Implement additionally a POST route to member genes.
